Created Material instances from FUEL data
FUEL blocks are defined differently than MATERIAL in Vera. I completed
the method Case.__get_fuel(fuel) in 'read_xml.py', to instantiate a
vanilla Material object for each fuel as defined in Vera.

This required the creation of a new module, 'isotopes.py', containing
the relative atomic weights of Oxygen and of nuclides I expect to
encounter in Vera decks.